+++ src/QueryFramework.Abstractions/GlobalUsings.cs
@@ -3,6 +3,8 @@ global using System.Collections;
 global using System.Collections.Generic;
 global using System.ComponentModel.DataAnnotations;
 global using System.Linq;
+global using System.Threading;
+global using System.Threading.Tasks;
 global using CrossCutting.Common;
 global using CrossCutting.Common.DataAnnotations;
 global using CrossCutting.Common.Extensions;