+++ src/QueryFramework.SqlServer/GlobalUsings.cs
@@ -4,6 +4,8 @@ global using System.Linq;
 global using System.Reflection;
 global using System.Runtime.CompilerServices;
 global using System.Text;
+global using System.Threading;
+global using System.Threading.Tasks;
 global using CrossCutting.Common.Extensions;
 global using CrossCutting.Data.Abstractions;
 global using CrossCutting.Data.Core;